flyttet alt ud i root mappen
