--- INFORMATION-BOARD-master-v2/README.md
@@ -1,2 +1,0 @@
-# INFORMATION-BOARD
-proejct
