--- INFORMATION-BOARD-master-v2/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "information-board",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "ejs": "^3.0.1",
-    "express": "^4.17.1",
-    "node-fetch": "^2.6.0",
-    "nodemon": "^2.0.2"
-  }
-}
