--- INFORMATION-BOARD-master-v2/views/pages/index.ejs
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <!-- Style -->
-    <link rel="stylesheet" href="../../assets/css/main.scss"> 
-    <title>Dashboard</title>
-</head>
-<body>
-    <!-- Nyheder -->
-    <%- include ('../partials/nyheder') -%>
-    <!-- Videos -->
-    <%- include ('../partials/video') -%>
-    <!-- Aktiviter -->
-    <%- include ('../partials/aktiviteter') -%>
-
-    <!-- Script -->
-    <script src="../../assets/js/main.js"></script>
-</body>
-</html>
