--- INFORMATION-BOARD-master-v2/views/partials/nyheder.ejs
@@ -1,9 +1,0 @@
-<h1><%- nyhedertitle -%></h1>
-<div>
-<% for (let i =0; i < news.length;  i++ ) { %>
-    <div>
-    <h2><%- news[i].vcTitle -%></h2>
-<p><%- news[i].txContent -%></p>
-</div>    
-    <% } %>
-</div>
