--- INFORMATION-BOARD-master-v2/views/partials/video.ejs
@@ -1,1 +1,0 @@
-<h1><%- videotitle -%></h1>
