--- nyversion/views/partials/aktiviteter.ejs
@@ -1,3 +1,0 @@
-<h1>Aktiviteter</h1>
-<h1><%- title -%></h1>
-<p><%- content -%></p>
