--- nyversion/views/partials/nyheder.ejs
@@ -1,3 +1,0 @@
-<h1>Nyheder</h1>
-<h1><%- title -%></h1>
-<p><%- content -%></p>
