--- nyversion/views/partials/video.ejs
@@ -1,3 +1,0 @@
-<h1>Video</h1>
-<h1><%- title -%></h1>
-<p><%- content -%></p>
