+++ views/pages/index.ejs
@@ -10,11 +10,11 @@
 </head>
 <body>
     <!-- Nyheder -->
-    <%- include ('../partials/nyheder') %>
+    <%- include ('../partials/nyheder') -%>
     <!-- Videos -->
-    <%- include ('../partials/video') %>
+    <%- include ('../partials/video') -%>
     <!-- Aktiviter -->
-    <%- include ('../partials/aktiviter') %>
+    <%- include ('../partials/aktiviteter') -%>
 
     <!-- Script -->
     <script src="../../assets/js/main.js"></script>
