--- views/partials/aktiviter.ejs
@@ -1,1 +1,0 @@
-<h1>Aktiviteter</h1>
