+++ views/partials/video.ejs
@@ -1,1 +1,1 @@
-<h1>Video</h1>
+<h1><%- videotitle -%></h1>
